refactor(st-09080): modularize multi-agent schemas
