+++ packages/patterns/package.json
@@ -19,8 +19,8 @@
   "scripts": {
     "build": "tsup src/index.ts --format esm,cjs --dts --clean",
     "dev": "tsup src/index.ts --format esm,cjs --dts --watch",
-    "test": "vitest",
-    "test:coverage": "vitest --coverage",
+    "test": "pnpm --dir ../.. exec vitest --config packages/patterns/vitest.config.ts",
+    "test:coverage": "pnpm --dir ../.. exec vitest --config packages/patterns/vitest.config.ts --coverage",
     "typecheck": "tsc -p tsconfig.typecheck.json --noEmit",
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
